embeddings/requirements.txt
colbert-ai==0.2.18
Flask==3.0.2
torch==2.2.0+cu118
transformers==4.37.2
